refactor: sf naming and aliases
