+++ src/commands/lightning/generate/app.ts
@@ -9,8 +9,8 @@ import { CreateOutput, LightningAppOptions } from '@salesforce/templates';
 import LightningAppGenerator from '@salesforce/templates/lib/generators/lightningAppGenerator';
 import { CreateUtil } from '@salesforce/templates/lib/utils';
 import { Messages } from '@salesforce/core';
-import { getCustomTemplates, runGenerator } from '../../../../utils/templateCommand';
-import { internalFlag, outputDirFlagLightning } from '../../../../utils/flags';
+import { getCustomTemplates, runGenerator } from '../../../utils/templateCommand';
+import { internalFlag, outputDirFlagLightning } from '../../../utils/flags';
 
 Messages.importMessagesDirectory(__dirname);
 const lightningCommonMessages = Messages.loadMessages('@salesforce/plugin-templates', 'lightning');
@@ -22,7 +22,8 @@ export default class LightningApp extends SfCommand<CreateOutput> {
   public static readonly summary = lightningCommonMessages.getMessage('summary', [BUNDLE_TYPE]);
   public static readonly description = lightningCommonMessages.getMessage('description', [BUNDLE_TYPE]);
   public static readonly examples = lightningAppMessages.getMessages('examples');
-
+  public static readonly aliases = ['force:lightning:app:create'];
+  public static readonly deprecateAliases = true;
   public static readonly flags = {
     name: Flags.string({
       char: 'n',
