+++ src/commands/visualforce/generate/component.ts
@@ -9,8 +9,8 @@ import { CreateOutput, VisualforceComponentOptions } from '@salesforce/templates
 import VisualforceComponentGenerator from '@salesforce/templates/lib/generators/visualforceComponentGenerator';
 import { CreateUtil } from '@salesforce/templates/lib/utils';
 import { Messages } from '@salesforce/core';
-import { outputDirFlag } from '../../../../utils/flags';
-import { runGenerator, getCustomTemplates } from '../../../../utils/templateCommand';
+import { outputDirFlag } from '../../../utils/flags';
+import { runGenerator, getCustomTemplates } from '../../../utils/templateCommand';
 
 const visualforceComponentFileSuffix = /.component$/;
 const VF_TYPE = 'Component';
@@ -23,7 +23,8 @@ export default class VisualforceComponent extends SfCommand<CreateOutput> {
   public static readonly summary = messages.getMessage('summary', [VF_TYPE]);
   public static readonly description = messages.getMessage('description', [VF_TYPE, VF_TYPE]);
   public static readonly examples = messages.getMessages('examples.component');
-
+  public static readonly aliases = ['force:visualforce:component:create'];
+  public static readonly deprecateAliases = true;
   public static readonly flags = {
     name: Flags.string({
       char: 'n',
@@ -59,7 +60,7 @@ export default class VisualforceComponent extends SfCommand<CreateOutput> {
       label: flags.label,
       template: 'DefaultVFComponent' as VisualforceComponentOptions['template'],
       outputdir: flags['output-dir'],
-      ...(typeof flags['api-version'] === 'string' ? { apiversion: flags['api-version'] } : {}),
+      apiversion: flags['api-version'],
     };
 
     return runGenerator({
